move ConnectionConfig to new module xknx.io.connection
